Experiment with QtLocation map
This is just an experimental commit to save the experimental state
and have it recorded somewhere. It'll be reverted with the next
commit.

I had to install

   qt5-plugin-geoservices-osm
   qt5-qtdeclarative-import-location
   qt5-qtlocation
   qt5-qtpositioning

in the emulator and

    qt5-qtpositioning-devel
    qt5-qtlocation-devel

in the mersdk vm to make this work.
(Log in into these VMs:
<https://sailfishos.org/develop/sdk-overview/develop-faq/#how_do_I_login_into_the_emulator_or_build_engine>)

For some weird reason, only QtLocation 5.0 and not 5.2 seems to be working.

Once we want to deploy this app as an RPM, these packages will have to be
added to the RPM building process:
<https://sailfishos.org/develop/sdk-overview/develop-packaging-apps/>

Also, the Nokia / here module does not work OOTB
--> you need a developer license (key?) to use it.

A mailing list entry about the QtLocation / QtPositioning modules and packages related
to sailfish:
<http://devel.sailfishos.narkive.com/tRGrEQl1/sailfishdevel-how-to-get-gps-location-in-qml-after-installing-qtpositioning-packages-to-targets-and->