Add PullDownMenu entry for a updating the sorting
// Github issue #13

Unfortunately, the fix for #14 didn't work properly:
Only calling sort(0) (via reSort()) did not have the desired effect,
because of this line in QSortFilterProxyModel:

>   if (d->dynamic_sortfilter && d->proxy_sort_column == column &&
>       d->sort_order == order)
>           return;

However, invoking the invalidate() slot does cause the proxy model
being sorted again.

But this seems to be doing too much: When one calls it onPositionChanged,
it completely resets the list and jumps to the top of the list.
Furthermore, the index is updated, too, and if this happens
while showing a GastroLocationDescription entry, suddenly another
item gets shown.

Hence, we have to find another solution for dynamic sorting.
In the meantime, we will stick with this PullDownMenu item.

Also, we update the translation, Github issue #17.