Automatically update sorting based on position
With [issue 9](https://github.com/Cutehacks/gel/issues/9)
fixed in Cutehacks::gel, we can now automatically
update the sorting when the current position information
was updated.

As this resorting causes the list to hang for a short
moment, we limit this update interval to a minimum
of 15 seconds.

The "Update sorting" entry will be removed for now.
Once Github issue #49 is implemented, we need it back
in the case that automatic sorting is disabled.

// Closes #45