Set a timer for downloading the JSON database
and request the JSON file on the file system when
the timer expires.

Because setting a timeout directly to the XMLHttpRequest
is not possible in the QML variant, we realized this
by converting the JsonDownloadHelper into an QML timer object.

TODO for this issue: Update the JSON file on the file system
when a new file was downloaded successfully.

// Github issue #4