Create a menu solution for both v-play and SFOS
This commit provides the BVApp concept of a NavigationMenu,
which can be assigned to an ApplicationWindow.

It includes the notion of an initialMenuItem, which is ignored
on Sailfish, as this item corresponds to the initialPage, which
we do not want to have in the pulley menu again as an item.
The initialPage in turn is ignored on v-play.

Currently you have to set both values (initialMenuItem, initialPage)
manually unfortunately. It should be possible to even generalize this.
But it is unclear how far we actually want to take the menu abstraction
anyway (or in which direction we want it to move, just consider toggleable
menu items for list view and map view, for example).
So we leave it like this atm.

Even v-play went with a "hacked" (i.e. non declarative, but imperative)
solution for the menu of their QtWorldSummit2016 app, when they
wanted to do something more sophisticated then the standard
NavigationMenu behavior:

    https://github.com/V-Play/QtWorldSummit2016-V-Play/blob/master/qml/QtWSMainItem.qml#L402

>   if(Theme.isAndroid) {
>     navigation.addNavigationItem(tracksNavItemComponent)
>     navigation.addNavigationItem(venueNavItemComponent)
> ...

This commit also turns every menuItem on v-play automatically
into an NavigationStack (for consitency with the typical stack based
navigation common on Sailfish).

BTW: The About page is still in rough state on we-pay ;)

// Github issue #67