[v-play] Provide a pushAttached() implementation
The pushAttached feature of the Sailfish pageStack
(which we use to put the map for one venue on the
right hand side of the VenueDescription page)
is not available on v-play.

A navigation paradigm like this (where the currently shown
page isn't the top of the navigation stack) is also very
uncommon on iOS and Android. We therefore extented the
pushAttached() interface by one parameter: An icon that
is shown on the right side of the navigation bar (where
you would have an 'arrow to the right' symbol if you
transferred the Sailfish style navigation scheme strictly),
which you can use to then actually call the attached-pushed
page.

In contrast to Sailfish, we do not actually push the page
when pushAttached() is being called. Instead the page is
only being pushed lazily once you hit the button. We needed
to adapt the way we supply the VenueMapPage with its
information for that purpose (which definitely can be
considered a code cleanup anyway).

Note: Like the AboutPage the VenueMapPage too is
in an unpolished state on v-play.

// Github issue #66