Bump version to 0.9
This is an internal maintenance release for enabling
the v-play port: Feature wise, the v-play variant is
on par with the Sailfish OS variant now.

As this an internal release, we do not bump the version
of the v-play variant now. We do however bump the
Sailfish version code to 0.90. (Yes, it should be 0.9,
but now that versions with wrong version codes of 0.81 etc. are
out in the wild, we have to wait until 1.0 to fix this...).