Fix #82: Straighten out menu situation
With the food and shopping menu entries in place,
on SailfishOS the menu behaves weirdly: The shopping list
gets pushed as a new page onto the stack of the food list
(without a 'PullDownMenu', which is still on the food list page).

We therefore create the entity of an 'ActionMenuItem',
which closely resembles Sailfish's PullDownMenu's 'MenuItem':
It has a 'clicked' signal, which you can connect to
for triggering some action when the menu item was clicked.

Currently, the 'ActionMenu' item uses the page component that
has been assigned to the 'initialPage' property of the
'ApplicationWindow'.

The 'NavigationMenu' abstraction on Sailfish is adapted in a
way that the menu automatically assigns itself to the first
page of Sailfish's global 'pageStack' (if it has a valid
'flickable' the 'PullDownMenu' can be attached to).