Merge shopping and food models VenueModels to one
(and with them their filter proxies) and have a proper
role, which defines the model category instead.

A nicer alternative would have been leaving
the VenueModels itself as they are and
instead have *one* filter model, which can handle both
source models.

But as Qt obviously does not provide a way of *easily*
implementing an aggregate proxy model as of now,

    http://blog.qt.io/blog/2012/01/12/qt-commercial-support-weekly-10-sorting-filtering-and-advanced-manipulation-with-proxy-models-2/
    https://stackoverflow.com/questions/14858629/can-i-create-a-summary-aggregate-model-using-qsortfilterproxymodel
    http://www.qtcentre.org/threads/23543-QSortFilterProxyModel-with-multiple-source-models

this is probably the nicest easy solution.

// Github issue #42