[v-play] fix map marker not pointing to correct position
When zooming in and out one can see, that the map marker icon does not
point to the correct position all the time. This is because, there is an
invisible rectangle around the icon that needs to be taken into
account.

With this patch we center the icon horizontally and align it to the
bottom in the parent. Additionally we set the anchor point when zooming
to that exact point.

The 'anchors.fill: parent' also fixes the icons in the IconToolBar:
there weren't vertically centered, but only horizontally.

Closes #88