[v-play] UX/UI: always return to first view
We find it bad UX, if a user puts views on the stack of menuItemA, then
switches to menuItemB, puts view on there and later comes back to
menuItemA and finds the old views there.

This also workarounds the problem that a user can open venues from
different views in the app and that our favorite status does not sync.

Closes #87