[v-play] Do not abuse Layout.alignment
Layout.alignment is for aligning an object in e.g. a RowLayout and not
the content of the object.

Default the AlignVCenter, where it is used currently in 4/5 places, and
set AlignBottom for the remaining IconButton explicitly.

Note: this is not needed on Sailfish. If one draws an opaque rectangle
around the icon, one will see, that the icon takes all the space of the
rectangle.

// github #131