Add a first map overview version
We had to change the way how the NavigationStackWithPushAttached works:
Now it puts the button directly to currentPage.rightBarItem. This makes
it possible to have several attached buttons on different pages
of the same navigation stack.

For SailfishOS, we added the 'pushed()' signal to the page abstraction and
implemented it via forwarding 'onPageContainerChanged()'.

// Github issue #7