Apply search to English description as well
With this patch also the English description is taken into account for a
search. Even if the system language is set to non-English. It works also
the other way around: the German description is searched, even if the
system language is set to English.

I think this behavior has only advantages, because it only produces more
matches and therefore potentially a positive result, e.g. the German
description for Goodies lacks the word "smoothie", but the English
description contains it, so if a non-English user searches for Smoothie,
he/she will get the correct result, because Goodies offers smoothies. :)

// github issue #27