[ios] Add possibility to choose map app
With this patch a user can choose if he/she wants to open the
destination either with Apple or with Google Maps.

If an app is not installed it will open the link in the browser.

On desktop one does not get a chooser presented. This behaviour is
implemented in NativeUtils.

Note: Currently the strings are hard-coded, because I don't see the
necessity for translating those strings for our supported languages.