[Android] Fix #262: Use Page instead of ListPage
When looking at the Android page push animation, which appears on
clicking a venue in the list, it appears like the list is shown again
in just one frame again after it was already faded out.

It seems that by using "Page" instead of "ListPage" this problem
can be alleviated.